conditional build for docker

+++ scripts/generate-release-notes.sh
@@ -2,13 +2,13 @@
 set -eo pipefail
 
 err_report() {
-    echo "Error on line $1"
+  echo "Error on line $1"
 }
 
 trap 'err_report $LINENO' ERR
 
 function usage() {
-    echo "Usage: $0 litentry|litmus path-to-output diff-tag release-type"
+  echo "Usage: $0 litentry|litmus path-to-output diff-tag release-type"
 }
 
 [ $# -ne 3 ] && [ $# -ne 4 ] && (usage; exit 1)
@@ -19,31 +19,34 @@ cd "$ROOTDIR"
 CHAIN=$1
 REPO=https://github.com/litentry/litentry-parachain
 
-# base image used to build the node binary
-NODE_BUILD_BASE_IMAGE=$(grep FROM docker/Dockerfile | head -n1 | sed 's/^FROM //;s/ as.*//')
+if [ "$3" != "runtime" ]; then
+  # base image used to build the node binary
+  NODE_BUILD_BASE_IMAGE=$(grep FROM docker/Dockerfile | head -n1 | sed 's/^FROM //;s/ as.*//')
 
-# somehow `docker inspect` doesn't pull our litentry-parachain image sometimes
-docker pull "$NODE_BUILD_BASE_IMAGE"
-docker pull "litentry/litentry-parachain:$RELEASE_TAG"
+  # somehow `docker inspect` doesn't pull our litentry-parachain image sometimes
+  docker pull "$NODE_BUILD_BASE_IMAGE"
+  docker pull "litentry/litentry-parachain:$RELEASE_TAG"
 
-NODE_VERSION=$(grep version node/Cargo.toml | head -n1 | sed "s/'$//;s/.*'//")
-NODE_BIN=litentry-collator
-NODE_SHA1SUM=$(shasum litentry-collator/"$NODE_BIN" | awk '{print $1}')
-if [ -f rust-toolchain.toml ]; then
-  NODE_RUSTC_VERSION=$(rustc --version)
-else
-  NODE_RUSTC_VERSION=$(docker run --rm "$NODE_BUILD_BASE_IMAGE" rustup default nightly 2>&1 | grep " installed" | sed 's/.*installed - //')
+  NODE_VERSION=$(grep version node/Cargo.toml | head -n1 | sed "s/'$//;s/.*'//")
+  NODE_BIN=litentry-collator
+  NODE_SHA1SUM=$(shasum litentry-collator/"$NODE_BIN" | awk '{print $1}')
+  if [ -f rust-toolchain.toml ]; then
+    NODE_RUSTC_VERSION=$(rustc --version)
+  else
+    NODE_RUSTC_VERSION=$(docker run --rm "$NODE_BUILD_BASE_IMAGE" rustup default nightly 2>&1 | grep " installed" | sed 's/.*installed - //')
+  fi
 fi
 
-SRTOOL_DIGEST_FILE=$CHAIN-parachain-runtime/$CHAIN-parachain-srtool-digest.json
-
-RUNTIME_VERSION=$(grep spec_version runtime/$CHAIN/src/lib.rs | sed 's/.*version: //;s/,//')
-RUNTIME_COMPRESSED_SIZE=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.size | sed 's/"//g')
-RUNTIME_RUSTC_VERSION=$(cat "$SRTOOL_DIGEST_FILE" | jq .rustc | sed 's/"//g')
-RUNTIME_COMPRESSED_SHA256=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.sha256 | sed 's/"//g')
-RUNTIME_COMPRESSED_BLAKE2=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.blake2_256 | sed 's/"//g')
-RUNTIME_COMPRESSED_SET_CODE_HASH=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.subwasm.proposal_hash | sed 's/"//g')
-RUNTIME_COMPRESSED_AUTHORIZE_UPGRADE_HASH=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.subwasm.parachain_authorize_upgrade_hash | sed 's/"//g')
+if [ "$3" != "client" ]; then
+  SRTOOL_DIGEST_FILE=$CHAIN-parachain-runtime/$CHAIN-parachain-srtool-digest.json
+  RUNTIME_VERSION=$(grep spec_version runtime/$CHAIN/src/lib.rs | sed 's/.*version: //;s/,//')
+  RUNTIME_COMPRESSED_SIZE=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.size | sed 's/"//g')
+  RUNTIME_RUSTC_VERSION=$(cat "$SRTOOL_DIGEST_FILE" | jq .rustc | sed 's/"//g')
+  RUNTIME_COMPRESSED_SHA256=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.sha256 | sed 's/"//g')
+  RUNTIME_COMPRESSED_BLAKE2=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.blake2_256 | sed 's/"//g')
+  RUNTIME_COMPRESSED_SET_CODE_HASH=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.subwasm.proposal_hash | sed 's/"//g')
+  RUNTIME_COMPRESSED_AUTHORIZE_UPGRADE_HASH=$(cat "$SRTOOL_DIGEST_FILE" | jq .runtimes.compressed.subwasm.parachain_authorize_upgrade_hash | sed 's/"//g')
+fi
 
 SUBSTRATE_DEP=$(grep sp-core node/Cargo.toml | sed 's/.*branch = "//;s/".*//')
 CUMULUS_DEP=$(grep cumulus-client-cli node/Cargo.toml | sed 's/.*branch = "//;s/".*//')
@@ -100,6 +103,11 @@ Cumulus                      : $CUMULUS_DEP
 EOF
 
 if [ "$GENESIS_RELEASE" = "true" ]; then
+  if [ "$3" = "runtime" ]; then
+    echo "genesis release requires to build client"
+    exit 1
+  fi
+
   GENESIS_STATE_HASH=$(shasum litentry-collator/$CHAIN-genesis-state | awk '{print $1}')
   GENESIS_WASM_HASH=$(shasum litentry-collator/$CHAIN-genesis-wasm | awk '{print $1}')
